requirements.txt
Django==1.9
Pillow==3.1.1
django-guardian==1.4.1
six==1.10.0
